Clean up and Organize Code Update on Imports
