+++ src/main/java/seedu/address/model/Model.java
@@ -3,23 +3,31 @@ package seedu.address.model;
 import java.util.function.Predicate;
 
 import javafx.collections.ObservableList;
-import seedu.address.model.person.Group;
+import seedu.address.model.group.Group;
 import seedu.address.model.person.Person;
 
 /**
  * The API of the Model component.
  */
 public interface Model {
-    /** {@code Predicate} that always evaluate to true */
+    /**
+     * {@code Predicate} that always evaluate to true
+     */
     Predicate<Person> PREDICATE_SHOW_ALL_PERSONS = unused -> true;
 
-    /** {@code Predicate} that always evaluate to true */
+    /**
+     * {@code Predicate} that always evaluate to true
+     */
     Predicate<Group> PREDICATE_SHOW_ALL_GROUPS = unused -> true;
 
-    /** Clears existing backing model and replaces with the provided new data. */
+    /**
+     * Clears existing backing model and replaces with the provided new data.
+     */
     void resetData(ReadOnlyAddressBook newData);
 
-    /** Returns the AddressBook */
+    /**
+     * Returns the AddressBook
+     */
     ReadOnlyAddressBook getAddressBook();
 
     /**
@@ -40,40 +48,24 @@ public interface Model {
     void addPerson(Person person);
 
     /**
-     * Returns true if a group with the same identity as {@code group} exists in the address book.
-     */
-    boolean hasGroup(Group group);
-
-    /**
-     * Adds the given group.
-     * {@code group} must not already exist in the address book.
-     */
-    void addGroup(Group group);
-
-    /**
      * Replaces the given person {@code target} with {@code editedPerson}.
      * {@code target} must exist in the address book.
      * The person identity of {@code editedPerson} must not be the same as another existing person in the address book.
      */
     void updatePerson(Person target, Person editedPerson);
 
-    /** Returns an unmodifiable view of the filtered person list */
+    /**
+     * Returns an unmodifiable view of the filtered person list
+     */
     ObservableList<Person> getFilteredPersonList();
-
-    /** Returns an unmodifiable view of the filtered group list */
-    ObservableList<Group> getFilteredGroupList();
 
     /**
      * Updates the filter of the filtered person list to filter by the given {@code predicate}.
+     *
      * @throws NullPointerException if {@code predicate} is null.
      */
     void updateFilteredPersonList(Predicate<Person> predicate);
 
-    /**
-     * Updates the filter of the filtered group list to filter by the given {@code predicate}.
-     * @throws NullPointerException if {@code predicate} is null.
-     */
-    void updateFilteredGroupList(Predicate<Group> predicate);
 
     /**
      * Returns true if the model has previous address book states to restore.
@@ -99,4 +91,28 @@ public interface Model {
      * Saves the current address book state for undo/redo.
      */
     void commitAddressBook();
+
+    /**
+     * Creates the given group.
+     * {@code createGroup} must not already exist in the system.
+     */
+    void createGroup(Group createGroup);
+
+    /**
+     * return true if there exist another group with the same name.
+     */
+    boolean hasGroup(Group checkGroup);
+
+    /**
+     * Returns an unmodifiable view of the filtered group list
+     */
+    ObservableList<Group> getFilteredGroupList();
+
+    /**
+     * Updates the filter of the filtered person list to filter by the given {@code predicate}.
+     *
+     * @throws NullPointerException if {@code predicate} is null.
+     */
+    void updateFilteredGroupList(Predicate<Group> predicate);
+
 }
