+++ src/main/java/seedu/address/model/UserPrefs.java
@@ -12,7 +12,7 @@ import seedu.address.commons.core.GuiSettings;
 public class UserPrefs {
 
     private GuiSettings guiSettings;
-    private Path addressBookFilePath = Paths.get("data" , "addressbook.xml");
+    private Path addressBookFilePath = Paths.get("data", "addressbook.xml");
 
     public UserPrefs() {
         setGuiSettings(500, 500, 0, 0);
